requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
python-dateutil>=2.8.2
drf-yasg>=1.21.7
djangorestframework-simplejwt>=5.3.0
